update order, fix bugs. Before deploy

+++ src/main/java/com/blankerdog/carService/dto/ItemDto.java
@@ -23,5 +23,5 @@ public class ItemDto {
     private Long orderId;
 
     @NotNull
-    private Map<String, Integer> itemPrice;
+    private Map<String, Float> itemPrice;
 }
